offload mailing to a child process
utilize a child process to handle I/O blocking operations specifically for sending emails. This approach frees up the main process to continue executing tasks without being blocked while waiting for the email to be sent

child process (process.mailer.utils.js)

This file listens for messages sent from the main process via process.on('message')
When a message containing email options is received, it creates a nodemailer transporter using the provided credentials.
It then sends the email using transporter.sendMail
After sending the email, the child process sends a message back to the main process indicating success or failure (process.send)

main process (mailer.utils.js)

The main process waits for a response from the child process using .on('message')
if the status from the child process is 'success', it resolves the promise in transportMail with the email response
If the status is 'failed', it rejects the promise with the error message

summary

By forking a child process, the main process is no longer blocked while the email is being sent. It can continue executing other tasks and receive a notification when the email operation is complete. This improves the responsiveness of the application.